refactor(start-server-core): use h3 barrel import
